requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# based-bot/main.py: 2,4
discord.py == 1.3.4

# based-bot/main.py: 5
google_cloud_secret_manager == 1.0.0
